lambda/requirements.txt
ask-sdk-core>=1.11.0
ask-sdk-s3-persistence-adapter>=1.0.0
python-dotenv>=0.15.0 